Implement updated song detector with variable capture tail and variable background calculation
This commit eliminates the direct decibel calculation, due to issues with normalization, and instead relies on a 10 minute moving average of power in the song spectrum. It uses a ratio of the current song power to the background song power, as well as the prior song to non-song power ratio, in order to trigger song capture.